Use hex colors (#ffffff) not 256-color names (color-255) for 24-bit compat
In emacs 28 on brew, 24 bit color is the default (even if on
screen-256color). Apparently, this causes color-* to be invalid.

Either hex colors or color names (like firebrick) should work, but to
use color names, we'd need to map them to our chosen hue first.
The easiest option is to use the corresponding hex color.

Note: base ANSI colors 0-15 are evidently rendered with 24-bit color
sequences as well, so terminal overrides for illegible colors like
'red' and 'blue' no longer work.